requirements.txt
pillow
gpxpy
